nginx-ssl-util: cleanup
